fix/organize code
